Add: Event and Requests API's

+++ ewm-service/src/main/java/ru/practicum/ewm/main/service/impl/CategoryServiceImpl.java
@@ -1,23 +1,29 @@
 package ru.practicum.ewm.main.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.ewm.main.exceptions.ConflictException;
+import ru.practicum.ewm.main.exceptions.IntegrityException;
 import ru.practicum.ewm.main.exceptions.NotFoundException;
 import ru.practicum.ewm.main.model.Category;
 import ru.practicum.ewm.main.model.dto.category.CategoryDto;
 import ru.practicum.ewm.main.model.dto.category.NewCategoryDto;
 import ru.practicum.ewm.main.model.mapper.CategoryMapper;
 import ru.practicum.ewm.main.repository.CategoryRepository;
+import ru.practicum.ewm.main.repository.EventRepository;
 import ru.practicum.ewm.main.service.CategoryService;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository repository;
+    private final EventRepository eventRepository;
 
     @Override
     @Transactional(readOnly = true)
@@ -40,27 +46,38 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     @Transactional
     public CategoryDto create(NewCategoryDto category) {
-        return CategoryMapper.toDto(repository.save(Category.builder()
-                .name(category.getName())
-                .build()));
+        try {
+            return CategoryMapper.toDto(repository.save(Category.builder()
+                    .name(category.getName())
+                    .build()));
+        } catch (DataIntegrityViolationException e) {
+            throw new IntegrityException(e.getMessage());
+        }
     }
 
     @Override
     @Transactional
     public CategoryDto update(NewCategoryDto dto, Long catId) {
-        Category category = repository.findById(catId)
-                .orElseThrow(() -> new NotFoundException("Category with id=%d was not found".formatted(catId)));
+        try {
+            Category category = repository.findById(catId)
+                    .orElseThrow(() -> new NotFoundException("Category with id=%d was not found".formatted(catId)));
 
-        if (!dto.getName().equals(category.getName())) {
-            category.setName(dto.getName());
+            if (!dto.getName().equals(category.getName())) {
+                category.setName(dto.getName());
+            }
+            return CategoryMapper.toDto(repository.save(category));
+        } catch (DataIntegrityViolationException e) {
+            throw new IntegrityException(e.getMessage());
         }
-
-        return CategoryMapper.toDto(repository.save(category));
     }
 
     @Override
     @Transactional
     public void deleteById(Long catId) {
+        if (eventRepository.existsByCategoryId(catId)) {
+            throw new ConflictException("The category is not empty");
+        }
+
         try {
             repository.deleteById(catId);
         } catch (EmptyResultDataAccessException ex) {
